feat: add (OneRenderBlockchainNotifications) to update (getBlockchainNotifications) hook
